+++ test/fixtures/webpackIgnore.html
@@ -73,12 +73,44 @@
 <!-- webpackIgnore: false -->
 <img src="image.png" />
 
+<!-- webpackIgnore: true -->
+<!-- webpackIgnore: false -->
+<!-- webpackIgnore: true -->
+<img src="image.png" />
+
 <!-- webpackIgnore:  -->
 <img src="image.png" />
 
 <!-- webpackIgnore: q -->
 <img src="image.png" />
 
+<!-- webpackIgnore:true -->
+<img src="image.png" />
+
+<!--    webpackIgnore:   true   -->
+<img src="image.png" />
+
+<!-- webpackIgnore: true --><img src="image.png" />
+<!-- webpackIgnore: false --><img src="image.png" />
+<!-- webpackIgnore: true --><!-- webpackIgnore: false --><img src="image.png" />
+<!-- webpackIgnore: false --><!-- webpackIgnore: true --><img src="image.png" />
+
+
+<!-- webpackIgnore: true -->
+<div></div>
+<img src="image.png" />
+
+<!-- webpackIgnore: true -->
+<br>
+<img src="image.png" />
+
+<!-- webpackIgnore: true -->
+<br />
+<img src="image.png" />
+
+<br />
+<!-- webpackIgnore: true -->
+<img src="image.png" />
 
 </body>
 </html>
